#25 Add startApp.js to use the express app return by app.js with specific config return by config.js
