+++ package.json
@@ -2,16 +2,16 @@
   "name": "PaperBadger",
   "version": "0.1.1",
   "description": "Exploring the use of digital badges for crediting contributors to scholarly papers for their work",
-  "main": "app.js",
+  "main": "index.js",
   "scripts": {
     "start": "npm-run-all --parallel server watch:js",
     "build:js": "webpack --config webpack.config.js --progress --profile --colors",
     "watch:js": "npm run build:js -- --watch",
     "clean": "rm -r public/js/*",
-    "server": "node web/app.js",
-    "jscs": "jscs -c node_modules/mofo-style/linters/.jscsrc web webpack.config.js",
-    "jshint": "jshint -c node_modules/mofo-style/linters/.jshintrc web webpack.config.js",
-    "jsbeautify": "js-beautify --config node_modules/mofo-style/linters/.jsbeautifyrc web/*.js webpack.config.js -r -n",
+    "server": "node src/index.js",
+    "jscs": "jscs -c node_modules/mofo-style/linters/.jscsrc src webpack.config.js",
+    "jshint": "jshint -c node_modules/mofo-style/linters/.jshintrc src webpack.config.js",
+    "jsbeautify": "js-beautify --config node_modules/mofo-style/linters/.jsbeautifyrc src/*.js webpack.config.js -r -n",
     "lint": "npm run jsbeautify && npm run jshint && npm run jscs",
     "build:production": "npm run build:js -- --optimize-minimize --optimize-dedupe",
     "start:production": "npm-run-all build:production server"
