+++ src/index.js
@@ -1,0 +1,7 @@
+var startApp = require('./startApp');
+
+var startAppCallback = function (app) {
+  console.log(app.name + ' started on ' + app.port);
+};
+
+startApp(startAppCallback);
